+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.6] - 2024-05-25
+
+### Fixed
+
+- Bug in `consistent!` for sparse matrices.
+
+## [0.4.5] - 2024-05-17
+
+### Fixed
+
+- Bug in `copy`.
+- Bug in sparse matrix-matrix products.
+- Performance improvements in `tuple_of_arrays`.
+
+### Added
+
+- Function `centralize` for sparse matrix.
+- `multicast` for arbitrary types.
+
 ## [0.4.4] - 2024-02-20
 
 ### Fixed